chore: delete redundancy
