+++ options/props/defaults.js
@@ -42,6 +42,26 @@ export const DEFAULT_SHUTTLE_ALL_ENDS = [
   "化学楼南往北",
   "云峰北侧",
 ];
+export const DEFAULT_QUERY_ALL_ENDS = [
+  "紫金港校区",
+  "玉泉校区",
+  "西溪校区",
+  "华家池校区",
+  "之江校区",
+  "工程师学院",
+  "信息港园区",
+  "水博园区",
+  "湖滨",
+  "紫金文苑",
+  "雅仕苑",
+  "丰谭路高教二期",
+  "南都德加东",
+  "城市心境",
+  "高教一期",
+  "恩济花园",
+  "高教二期",
+  "丰谭中学",
+];
 export const NOP = () => {};
 export const DEFAULT_ROUTE = {
   zIndex: 4,
