+++ pages/search/search.js
@@ -1,7 +1,6 @@
-import { fmtQueryResult, fmtQueryArrayResult } from "/util/fmtUnit";
-import { isObjectValueEqual } from "/util/queryhelper";
-import { dataAlert } from "/util/data";
 import { load } from "/util/cache";
+import { fmtQueryArrayResult, fmtQueryResult, isObjectValueEqual } from "/util/formatter";
+import { alertData } from "/util/notification";
 
 Page({
   data: {
@@ -107,6 +106,6 @@ Page({
         initStart: query.start,
       });
     }
-    !load("dataAlert").data && dataAlert();
+    !load("dataAlert").data && alertData();
   },
 });
